pyproject.toml
[project]
name = "playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jonatan Snyders", email = "jonatansnyders@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "pytest-xprocess>=1.0.2",
]

[project.scripts]
playground = "playground:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
